back to working as well as it did before I messed with it
