+++ .gitignore
@@ -1,1 +1,3 @@
 debug_outputs/*
+__pycache__/*
+*/__pycache__/*
